Clean up warnings from IntelliJ

+++ WikipediaCleaner/src/org/wikipediacleaner/api/check/algorithm/a0xx/a04x/a042/CheckErrorAlgorithm042.java
@@ -45,7 +45,7 @@ public class CheckErrorAlgorithm042 extends CheckErrorAlgorithmBase {
       return false;
     }
     Integer ns = analysis.getPage().getNamespace();
-    if ((ns == null) || (ns.intValue() != Namespace.MAIN)) {
+    if ((ns == null) || (ns != Namespace.MAIN)) {
       return false;
     }
