+++ WikipediaCleaner/src/org/wikipediacleaner/gui/swing/basic/TaskConfirmDialog.java
@@ -62,7 +62,7 @@ public class TaskConfirmDialog implements Runnable {
    */
   public int getResult() {
     if (result != null) {
-      return result.intValue();
+      return result;
     }
     return JOptionPane.CLOSED_OPTION;
   }
